+++ src/__init__.py
@@ -1,1 +1,31 @@
 """Package: drug-interaction-db"""
+
+from src.cyp450_tagger import (
+    KNOWN_ENZYMES,
+    EnzymeTag,
+    extract_enzymes,
+    filter_by_enzyme,
+    summarise_by_enzyme,
+    tag_interactions,
+)
+from src.main import DrugInteractionDB
+from src.polypharmacy_risk_scorer import (
+    PairScore,
+    RiskReport,
+    report_to_dataframe,
+    score_regimen,
+)
+
+__all__ = [
+    "DrugInteractionDB",
+    "EnzymeTag",
+    "KNOWN_ENZYMES",
+    "PairScore",
+    "RiskReport",
+    "extract_enzymes",
+    "filter_by_enzyme",
+    "report_to_dataframe",
+    "score_regimen",
+    "summarise_by_enzyme",
+    "tag_interactions",
+]